Move the 'updates' block to its own module to ease testing
