+++ updates.h
@@ -1,0 +1,6 @@
+#ifndef EMS_UPDATES_H
+#define EMS_UPDATES_H
+
+int apply_updates(int page, int verbose, struct updates *updates);
+
+#endif /* EMS_UPDATES_H */
